Added listing pages

+++ database/migrations/2023_12_07_175940_add_fields_to_listings_table.php
@@ -16,7 +16,7 @@ class AddFieldsToListingsTable extends Migration
         Schema::table('listings', function (Blueprint $table) {
            $table->unsignedTinyInteger('beds');
            $table->unsignedTinyInteger('baths');
-           $table->unsignedTinyInteger('area');
+           $table->unsignedSmallInteger('area');
 
            $table->tinyText('city');
            $table->tinyText('code');
